Move all time utilities to a separate class
